refactor: move sut to a factory helper method
